Move utils to io.rs
