+++ core/primitives/src/stateless_validation.rs
@@ -1,12 +1,11 @@
 use std::collections::{HashMap, HashSet};
-use std::io::Write;
-use std::io::{self, Read};
 
 use crate::challenge::PartialState;
 use crate::congestion_info::CongestionInfo;
 use crate::sharding::{ChunkHash, ReceiptProof, ShardChunkHeader, ShardChunkHeaderV3};
 use crate::transaction::SignedTransaction;
 use crate::types::EpochId;
+use crate::utils::io::{CountingRead, CountingWrite};
 use crate::validator_signer::{EmptyValidatorSigner, ValidatorSigner};
 use borsh::{BorshDeserialize, BorshSerialize};
 use bytes::{Buf, BufMut};
@@ -14,6 +13,7 @@ use near_crypto::{PublicKey, Signature};
 use near_primitives_core::hash::CryptoHash;
 use near_primitives_core::types::{AccountId, Balance, BlockHeight, ShardId};
 use near_primitives_core::version::PROTOCOL_VERSION;
+use std::io::Read;
 
 /// An arbitrary static string to make sure that this struct cannot be
 /// serialized to look identical to another serialized struct. For chunk
@@ -168,95 +168,6 @@ impl EncodedChunkStateWitness {
 
     pub fn as_slice(&self) -> &[u8] {
         &self.0
-    }
-}
-
-/// Wrapper for Write that counts number of bytes written.
-/// It also allow setting an optional limit to stop writing.
-struct CountingWrite<W: Write> {
-    inner: W,
-    /// Total number of bytes written.
-    written: u64,
-    /// If set, the number of bytes allowed to be written.
-    /// If this limit is reached, any additional write will return an error.
-    limit: Option<u64>,
-}
-
-impl<W: Write> CountingWrite<W> {
-    fn new_with_limit(inner: W, limit: bytesize::ByteSize) -> Self {
-        Self { inner, written: 0, limit: Some(limit.as_u64()) }
-    }
-
-    fn new(inner: W) -> Self {
-        Self { inner, written: 0, limit: None }
-    }
-
-    fn bytes_written(&self) -> bytesize::ByteSize {
-        bytesize::ByteSize::b(self.written)
-    }
-
-    fn into_inner(self) -> W {
-        self.inner
-    }
-}
-
-impl<W: Write> Write for CountingWrite<W> {
-    fn write(&mut self, buffer: &[u8]) -> io::Result<usize> {
-        if let Some(limit) = self.limit {
-            if self.written.saturating_add(buffer.len() as u64) >= limit {
-                return Err(std::io::Error::other(format!("Reached the limit of {} bytes", limit)));
-            }
-        }
-        let last_written = self.inner.write(buffer)?;
-        self.written = self.written.saturating_add(last_written as u64);
-        Ok(last_written)
-    }
-
-    fn flush(&mut self) -> io::Result<()> {
-        self.inner.flush()
-    }
-}
-
-/// Wrapper for Write that counts number of bytes written.
-/// It also allow setting an optional limit to stop writing.
-struct CountingRead<R: Read> {
-    inner: R,
-    /// Total number of bytes read.
-    read: u64,
-    /// If set, the number of bytes allowed to be written.
-    /// If this limit is reached, any additional write will return an error.
-    limit: Option<u64>,
-}
-
-impl<R: Read> CountingRead<R> {
-    fn new_with_limit(inner: R, limit: bytesize::ByteSize) -> Self {
-        Self { inner, read: 0, limit: Some(limit.as_u64()) }
-    }
-
-    fn new(inner: R) -> Self {
-        Self { inner, read: 0, limit: None }
-    }
-
-    fn bytes_read(&self) -> bytesize::ByteSize {
-        bytesize::ByteSize::b(self.read)
-    }
-
-    fn into_inner(self) -> R {
-        self.inner
-    }
-}
-
-impl<R: Read> Read for CountingRead<R> {
-    fn read(&mut self, buffer: &mut [u8]) -> io::Result<usize> {
-        let buffer_size = buffer.len();
-        if let Some(limit) = self.limit {
-            if self.read.saturating_add(buffer_size as u64) >= limit {
-                return Err(std::io::Error::other(format!("Reached the limit of {} bytes", limit)));
-            }
-        }
-        let last_read = self.inner.read(buffer)?;
-        self.read = self.read.saturating_add(last_read as u64);
-        Ok(last_read)
     }
 }
